Extracts spare part field to a dedicated component
Refactors the spare part selection logic, including data fetching and `Autocomplete` rendering, into a new `SparePartFormikField` component.

This improves modularity and reusability, centralizing the complex field logic and making the `SparePartsArrayField` component cleaner and easier to read. It also introduces optimized SWR caching strategies for spare part data.
